Fixed A Bug, As Well As Updated To DFU v1.0.0, Due To Problem, 3/20/2024
Fixed A Bug, As Well As Updated To DFU v1.0.0, Due To Problem. Seems in DFU v0.15.1 the "InputManager.cs" class was changed slightly so the method I was using to detect keyboard inputs would break in the precompiled mod, so decided it was best to just update the build in DFU v1.0.0, 3/20/2024.